--- client/src/pages/Home/components/index.ts
@@ -1,6 +1,0 @@
-import Card from './Card';
-import CardPrice from '../../../../unusedCode/CardPrice';
-import CardsContainer from './CardsContainer';
-import InputSearch from './InputSearch';
-
-export { CardsContainer, Card, CardPrice, InputSearch };
