+++ client/src/utils/Api/Api.ts
@@ -1,8 +1,8 @@
 import axios from 'axios';
 import { getLocationWhenEmpty } from '../utils';
 import { ApiKeys } from '../constants';
-import { CardItem } from '../../pages/Home/components/Card';
 import { UnsplashPhotoSearch, SearchResponse, RandomResponse } from './types';
+import { CardItem } from '../../components/Card';
 
 export const unsplashApi = axios.create({
   baseURL: 'https://api.unsplash.com/',
@@ -10,39 +10,6 @@ export const unsplashApi = axios.create({
     Authorization: ApiKeys.first,
   },
 });
-
-// const unsplashRandomMapping = (data: RandomResponse): CardItem => {
-//   const { id, urls, created_at, location, user } = data;
-
-//   return {
-//     id,
-//     name: user.name || location?.name || user.location || getLocationWhenEmpty(),
-//     image: urls.regular,
-//     releaseDate: created_at,
-//   };
-// };
-
-// const unsplashSearchMapping = (data: UnsplashPhotoSearch): CardItem => {
-//   const { id, urls, created_at, user } = data;
-//   return {
-//     id,
-//     name: user.name || user.location || getLocationWhenEmpty(),
-//     image: urls.regular,
-//     releaseDate: created_at,
-//   };
-// };
-
-// const getRandom = async <Data>(url: Url<Data>, params: Params<Data>) => {
-//   const { data } = await unsplashApi<Data[]>(url, { params });
-//   return data;
-// };
-
-// const getSearch = async <Data>(url: Url<Data>, params: Params<Data>) => {
-//   const { data } = await unsplashApi<Data[]>(url, { params });
-//   return data;
-// };
-// `photos/random`
-// `search/photos`
 
 const unsplashMapping = (data: UnsplashPhotoSearch): CardItem => {
   const { id, urls, created_at, user } = data;
