+++ client/tsconfig.json
@@ -18,11 +18,5 @@
     "noImplicitAny": true,
     "composite": true
   },
-  "include": [
-    "src",
-    ".eslintrc.cjs",
-    "vite.config.ts",
-    "unusedCode/CardPrice.tsx",
-    "unusedCode/hooks.tsx"
-  ]
+  "include": ["src", ".eslintrc.cjs", "vite.config.ts", "unusedCode/*/*"]
 }
